feat: Move dialog in their own folder
